refactor(map): extract pin-sprite pipeline into shared module
Hoist the duplicated MapLibre pin-sprite pipeline out of /map-next,
AreaMapNext, and MultiPlaceMapNext into a single $lib/map/maplibreSprites
module. The /map-next cascade-fallback fetchIconInnerSvg becomes the
canonical version, so AreaMapNext and MultiPlaceMapNext now also fall back
through material-symbols and currency-bitcoin instead of failing on the
~6 icon names that don't exist in ic:outline. The sprite-promise cache is
per-map (WeakMap-keyed) so multiple maps co-existing in one session don't
cross-pollute their image registries, and resolved promises evict from the
cache so a subsequent setStyle()-driven image clear regenerates lazily.
