+++ src/components/MultiPlaceMapNext.svelte
@@ -10,6 +10,10 @@ import { onDestroy, onMount } from "svelte";
 
 import AreaMerchantDrawer from "$components/area/AreaMerchantDrawer.svelte";
 import MapLoadingEmbed from "$components/MapLoadingEmbed.svelte";
+import {
+	ensureSprite,
+	installPlaceholderHandler,
+} from "$lib/map/maplibreSprites";
 import { theme } from "$lib/theme";
 import type { SavedPlace } from "$lib/types";
 
@@ -55,91 +59,9 @@ let mapLoaded = false;
 let styleLoaded = false;
 let lastAppliedTheme: "light" | "dark" | undefined;
 
-// Mirrors Icon.svelte's material-symbol resolution. Keep in sync with /map-next.
-const materialExceptions: Record<string, string> = {
-	camping: "material-symbols:camping-rounded",
-	gate: "material-symbols:gate",
-	cooking: "material-symbols:cooking",
-	dentistry: "material-symbols:dentistry",
-	sauna: "material-symbols:sauna",
-	info_outline: "material-symbols:info-outline",
-	skull: "material-symbols:skull",
-	currency_bitcoin: "material-symbols:currency-bitcoin",
-};
-
-const resolveIconifyName = (icon: string): string => {
-	const key = icon === "question_mark" ? "currency_bitcoin" : icon;
-	return materialExceptions[key] ?? `ic:outline-${key.replace(/_/g, "-")}`;
-};
-
-const PIN_PATH =
-	"M0 16.0333C0 6.08 8.05161 0.131836 15.8361 0.131836C23.6205 0.131836 31.6721 6.08 31.6721 16.0333C31.6721 26.461 16.9494 41.3035 16.3229 41.9301C16.1941 42.0595 16.0185 42.1318 15.8361 42.1318C15.6536 42.1318 15.478 42.0595 15.3493 41.9301C14.7227 41.3035 0 26.461 0 16.0333Z";
-
-const PIN_FILL_REGULAR = "#0E95AF";
-
 // Saved places are always rendered with the bitcoin icon — mirrors the legacy
 // Leaflet MultiPlaceMap which hardcodes generateIcon(..., "currency_bitcoin").
 const SAVED_PLACE_ICON = "currency_bitcoin";
-
-const spriteName = (icon: string): string => `pin-r-${icon}`;
-
-// Per-instance cache so a setStyle() re-init can clear and regenerate cleanly.
-let spritePromises = new Map<string, Promise<void>>();
-
-const fetchIconInnerSvg = async (icon: string): Promise<string> => {
-	const iconifyName = resolveIconifyName(icon);
-	const path = iconifyName.replace(":", "/");
-	const url = `https://api.iconify.design/${path}.svg?color=white&width=20&height=20`;
-	const res = await fetch(url);
-	if (!res.ok) throw new Error(`Iconify fetch failed: ${res.status} ${url}`);
-	return await res.text();
-};
-
-const buildCompositeSvg = (innerSvg: string): string => {
-	// innerSvg is a complete <svg>...</svg> document; nesting an SVG inside an
-	// outer SVG is valid and rasterizes correctly through <img>.
-	return `<svg xmlns="http://www.w3.org/2000/svg" width="32" height="43" viewBox="0 0 32 43"><path d="${PIN_PATH}" fill="${PIN_FILL_REGULAR}"/><g transform="translate(6, 5.75)">${innerSvg}</g></svg>`;
-};
-
-const loadSvgImage = (svg: string): Promise<HTMLImageElement> =>
-	new Promise((resolve, reject) => {
-		const img = new Image();
-		img.crossOrigin = "anonymous";
-		img.onload = () => resolve(img);
-		img.onerror = (err) => reject(err);
-		img.src = `data:image/svg+xml;utf8,${encodeURIComponent(svg)}`;
-	});
-
-const ensureSprite = (m: MapLibreMap, icon: string): Promise<void> => {
-	const key = spriteName(icon);
-	if (m.hasImage(key)) return Promise.resolve();
-	const existing = spritePromises.get(key);
-	if (existing) return existing;
-	const promise = (async () => {
-		const inner = await fetchIconInnerSvg(icon);
-		const composite = buildCompositeSvg(inner);
-		const img = await loadSvgImage(composite);
-		if (!m.hasImage(key)) m.addImage(key, img, { pixelRatio: 1 });
-		m.triggerRepaint();
-	})();
-	spritePromises.set(key, promise);
-	promise.catch(() => spritePromises.delete(key));
-	return promise;
-};
-
-// 1×1 transparent placeholder so styleimagemissing doesn't spam warnings before
-// composite sprites resolve. Once the real sprite is added via addImage(), it
-// replaces this stub.
-const transparentPixel = (): ImageData => {
-	const canvas = document.createElement("canvas");
-	canvas.width = 1;
-	canvas.height = 1;
-	const ctx = canvas.getContext("2d");
-	if (!ctx) {
-		return new ImageData(new Uint8ClampedArray([0, 0, 0, 0]), 1, 1);
-	}
-	return ctx.getImageData(0, 0, 1, 1);
-};
 
 const buildFeatureCollection = (
 	list: SavedPlace[],
@@ -209,7 +131,7 @@ const syncPlacesSource = (m: MapLibreMap, list: SavedPlace[]) => {
 	const source = m.getSource("places") as GeoJSONSource | undefined;
 	if (!source) return;
 	source.setData(buildFeatureCollection(list));
-	ensureSprite(m, SAVED_PLACE_ICON);
+	ensureSprite(m, SAVED_PLACE_ICON, false);
 };
 
 // Register sprites + sources + layers once per style. Called from both the
@@ -323,12 +245,7 @@ const initializeMap = async () => {
 	});
 	map.addControl(geolocate, "top-right");
 
-	map.on("styleimagemissing", (event) => {
-		if (!map) return;
-		const id = event.id;
-		if (!id || map.hasImage(id)) return;
-		map.addImage(id, transparentPixel(), { pixelRatio: 1 });
-	});
+	installPlaceholderHandler(map);
 
 	map.on("load", () => {
 		if (!map) return;
@@ -340,15 +257,15 @@ const initializeMap = async () => {
 };
 
 // Theme reactivity — swap the basemap, then re-register places overlay on the
-// resulting style.load event. Sprite cache is cleared so the icon is
-// regenerated lazily on first render against the new style.
+// resulting style.load event. The shared sprite cache evicts resolved promises
+// on completion, so any sprites that didn't survive the style swap regenerate
+// lazily on first render against the new style.
 const applyTheme = (next: "light" | "dark" | undefined) => {
 	if (!map) return;
 	if (!styleLoaded) return;
 	if (next === lastAppliedTheme) return;
 	lastAppliedTheme = next;
 	styleLoaded = false;
-	spritePromises = new Map();
 	const onStyleLoad = () => {
 		if (!map) return;
 		initializeMapContents(map);
